requirements.txt
asgiref==3.4.1
coverage==5.5
Django==3.2.6
django-isbn-field==0.5.2
psycopg2-binary==2.9.1
python-stdnum==1.16
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
unicorn==1.0.3
